split out priority queue as a new file. the element in queue is still not ordered correctly.
