+++ cpp/include/reneu/segmentation/dendrogram.hpp
@@ -11,7 +11,7 @@
 
 #include <xtensor/xtensor.hpp>
 #include "disjoint_sets.hpp"
-#include "reneu/type_aliase.hpp"
+#include "../type_aliase.hpp"
 
 
 namespace reneu{
